requirements.txt
absl-py>=0.13.0
matplotlib>=3.4.2
numpy>=1.21.0
pandas>=1.2.5
scipy>=1.7.0
